requirements.txt
# === Core ML Libraries ===
scikit-learn==1.6.1
xgboost==2.0.3
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1

# === Web Framework ===
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0

# === Optional ===
matplotlib==3.8.4
seaborn==0.13.2
